cli: extract shared rewrite engine into rewrite/ module
Move the file-rewriting engine out of the filter command into a dedicated
crate::rewrite module, so filter/sort/compress/decompress are thin adapters
over one pipeline instead of reaching into commands::filter.

- rewrite/mod.rs: options, the metadata -> messages -> attachments pipeline,
  and writer setup.
- rewrite/message_source.rs: MessageSource unifies the previously separate
  indexed and linear message loops, plus summary synthesis and helpers.
- rewrite/aux_records.rs: metadata/attachment copy (index-or-scan), now shared
  by both the indexed and linear paths.

No behavior change: the engine takes a single input, and the structure leaves
room for a future multi-input merge to compose several MessageSources.
