+++ rust/cli/src/commands/compress.rs
@@ -1,8 +1,8 @@
 use anyhow::Result;
 
 use crate::cli::CompressCommand;
-use crate::commands::filter::{self, TranscodeCommandOptions};
 use crate::context::CommandContext;
+use crate::rewrite::{self, TranscodeCommandOptions};
 
 pub fn run(ctx: &CommandContext, args: CompressCommand) -> Result<()> {
     let options = TranscodeCommandOptions::new(args.file, args.output, args.chunk_size)
@@ -10,7 +10,7 @@ pub fn run(ctx: &CommandContext, args: CompressCommand) -> Result<()> {
         .use_chunks(true)
         .include_metadata(true)
         .include_attachments(true);
-    filter::run_transcode(
+    rewrite::run(
         options,
         crate::source::SourceOptions::new(ctx.allow_remote_scan()),
     )
