+++ rust/cli/src/commands/decompress.rs
@@ -1,8 +1,8 @@
 use anyhow::Result;
 
 use crate::cli::DecompressCommand;
-use crate::commands::filter::{self, TranscodeCommandOptions};
 use crate::context::CommandContext;
+use crate::rewrite::{self, TranscodeCommandOptions};
 
 pub fn run(ctx: &CommandContext, args: DecompressCommand) -> Result<()> {
     let options = TranscodeCommandOptions::new(args.file, args.output, args.chunk_size)
@@ -10,7 +10,7 @@ pub fn run(ctx: &CommandContext, args: DecompressCommand) -> Result<()> {
         .include_metadata(true)
         .include_attachments(true)
         .use_chunks(true);
-    filter::run_transcode(
+    rewrite::run(
         options,
         crate::source::SourceOptions::new(ctx.allow_remote_scan()),
     )
